chore: move circle style to utils
