+++ src/components/molecules/CollectionScroll.tsx
@@ -1,4 +1,4 @@
-import React, {useCallback, useMemo} from 'react';
+import React, {useCallback} from 'react';
 import {AtomView} from '../atoms/AtomView';
 import {AtomText} from '../atoms/AtomText';
 import {
@@ -7,15 +7,11 @@ import {
   ListRenderItemInfo,
   StyleSheet,
 } from 'react-native';
-import {
-  DEFAULT_SCROLL_VIEW_PROPS,
-  WINDOW_HEIGHT,
-  WINDOW_WIDTH,
-} from '../../styles/common';
+import {DEFAULT_SCROLL_VIEW_PROPS} from '../../styles/common';
 import {useNavigation} from '@react-navigation/native';
-import {roundOff} from '../../lib/utils';
 import {Category} from '../../types/apiDataTypes';
 import {RootNavigation} from '../../types/navTypes';
+import {getCircleStyle} from '../../styles/utils';
 
 type CollectionScrollProps = {
   collections: Category[];
@@ -33,15 +29,6 @@ const ItemSeparatorComponent = () => <AtomView pR="xs" />;
 const CollectionScroll = ({collections}: CollectionScrollProps) => {
   const navigation = useNavigation<RootNavigation>();
 
-  // responsive circle style
-  const getCircleStyle = useMemo(() => {
-    const dim = WINDOW_WIDTH > WINDOW_HEIGHT ? WINDOW_WIDTH : WINDOW_HEIGHT;
-    const width = roundOff(dim * 0.16);
-    const height = roundOff(dim * 0.16);
-    const borderRadius = roundOff(dim / 2);
-    return {width, height, borderRadius};
-  }, []);
-
   const renderCollection: ListRenderItem<Category> = useCallback(
     ({item}: ListRenderItemInfo<Category>) => {
       const onPressCategory = () =>
@@ -54,7 +41,7 @@ const CollectionScroll = ({collections}: CollectionScrollProps) => {
           justifyContent="center"
           alignItems="center"
           backgroundColor="pineGreen"
-          style={getCircleStyle}>
+          style={getCircleStyle()}>
           <AtomText
             color="white"
             fontWeight={'bold'}
@@ -66,7 +53,7 @@ const CollectionScroll = ({collections}: CollectionScrollProps) => {
         </AtomView>
       );
     },
-    [navigation, getCircleStyle],
+    [navigation],
   );
 
   return (
